feat: Enhance Iceberg filter pushdown to include desired columns in query execution
