+++ presto-native-execution/presto_cpp/main/connectors/IcebergPrestoToVeloxConnector.cpp
@@ -515,6 +515,25 @@ IcebergPrestoToVeloxConnector::toVeloxTableHandle(
     }
   }
 
+  // Row-lineage columns (_row_id, _last_updated_sequence_number) requested
+  // purely in the SELECT projection -- not in a filter predicate or as a
+  // partition column -- must also reach finalDataColumns below. Otherwise the
+  // Velox reader never learns their Parquet field IDs and silently falls back
+  // to firstRowId+position / the file's own dataSequenceNumber instead of
+  // each row's real physical value written by an external row-preserving
+  // UPDATE/MERGE.
+  if (icebergLayout->requestedColumns) {
+    for (const auto& entry : *icebergLayout->requestedColumns) {
+      if (kRowLineageColumnNames.count(entry.columnIdentity.name) &&
+          columnNames.emplace(entry.columnIdentity.name).second) {
+        columnHandles.emplace_back(
+            std::dynamic_pointer_cast<
+                const velox::connector::hive::HiveColumnHandle>(
+                std::shared_ptr(toVeloxColumnHandle(&entry, typeParser))));
+      }
+    }
+  }
+
   auto icebergTableHandle =
       std::dynamic_pointer_cast<const protocol::iceberg::IcebergTableHandle>(
           tableHandle.connectorHandle);
